License fix: Move sparse utility functions to common
