+++ docs/source/reference/sparse.rst
@@ -1,0 +1,6 @@
+Sparse array utils
+------------------
+
+.. automodule:: libertem.common.sparse
+   :members:
+   :undoc-members:
