+++ tests/test_smoke.py
@@ -16,6 +16,15 @@ import libertem.preload
 import libertem.udf
 import libertem.analysis.gridmatching   # NOQA: F401
 
+# old imports, refs #1031
+from libertem.corrections import CorrectionSet  # NOQA: F401
+from libertem.corrections.corrset import CorrectionSet as _  # NOQA: F401
+from libertem.corrections.detector import (  # NOQA: F401
+    correct, RepairDescriptor, RepairValueError, correct_dot_masks
+)
+
+from libertem.masks import to_dense, to_sparse, is_sparse  # NOQA: F401
+
 
 def test_stuff():
     print("stuff")
